% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.dat.beta}
\alias{assess.dat.beta}
\title{Assess estimation of eta when X and C have beta distributions}
\usage{
assess.dat.beta(n, q, B, s2, x.thetas, x.gamma, c.gamma, x.correct, c.correct)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{x.gamma}{a positive number, gamma parameter for X}

\item{c.gamma}{a positive number, gamma parameter for C}

\item{x.correct}{logical, an indicator of whether X|Z should be correctly
specified as conditional beta as opposed to incorrectly specified as marginal
beta}

\item{c.correct}{logical, an indicator of whether C|Z should be correctly
specified as conditional beta as opposed to incorrectly specified as marginal
beta}

\item{x.mean}{a positive number, the mean of X}

\item{n.rep}{a positive integer, the number of simulated replicates}
}
\value{
plot of empirical distribution of estimated parameters for eta1 and
eta2
}
\description{
Assess estimation of eta when X and C have beta distributions
}
