% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.dat.gamma}
\alias{assess.dat.gamma}
\title{Assess data generation with X|Z and C|Z having gamma distributions}
\usage{
assess.dat.gamma(
  n,
  q,
  B,
  s2,
  x.means,
  x.shape,
  c.shape,
  specify.x.gamma,
  specify.c.gamma
)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}

\item{x.mean}{a positive number, the mean of X}
}
\value{
a plot of generated data
}
\description{
Assess data generation with X|Z and C|Z having gamma distributions
}
