% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.dat}
\alias{assess.dat}
\title{Assess data generation}
\usage{
assess.dat(n, q, B, x.mean, x.shape, c.shape)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{x.mean}{a positive number, the mean of X}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}
}
\value{
a plot of generated data
}
\description{
Assess data generation
}
