% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.ee}
\alias{assess.ee}
\title{Assess estimating equation}
\usage{
assess.ee(ee, digits = 2)
}
\arguments{
\item{ee}{a numeric matrix, estimating function values at each observation}

\item{digits}{a non-negative integer, number of digits to round results,
default is 2}
}
\value{
individual or summation estimating function values
}
\description{
Assess estimating equation
}
