% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_variance_helpers.R
\name{d.log.fxz}
\alias{d.log.fxz}
\title{derivative of log-likelihood assuming X|Z ~ Beta}
\usage{
d.log.fxz(dat, theta, args = list(), return.sums = T)
}
\arguments{
\item{dat}{a data frame including columns Y, W, Delta, Z}

\item{theta}{a numeric vector, beta parameters for X|Z}

\item{args}{list of additional arguments}

\item{return.sums}{logical indicator for returning sum of scores as opposed
to individual scores, default is TRUE}
}
\value{
a matrix of estimating equation values
}
\description{
derivative of log-likelihood assuming X|Z ~ Beta
}
