% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_fit.R
\name{g.softmax}
\alias{g.softmax}
\title{Derivative of softmax transformation}
\usage{
g.softmax(x, wts)
}
\arguments{
\item{x}{a numeric vector, values to be transformed}

\item{wts}{a numeric vector, weights}
}
\value{
a numeric matrix
}
\description{
Derivative of softmax transformation
}
