% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_gamma.R
\name{gen.data.1}
\alias{gen.data.1}
\title{generate data with X and C having gamma distributions}
\usage{
gen.data.1(n, q, B, s2, x.means, x.shape, c.shape)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}
}
\value{
a data frame
}
\description{
generate data with X and C having gamma distributions
}
