% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_gamma.R
\name{gen.data.gamma}
\alias{gen.data.gamma}
\title{generate data with X and C having gamma distributions}
\usage{
gen.data.gamma(n, q, B, s2, x.means, x.shape, c.shape)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}
}
\value{
a list of the following data frames:
\itemize{
\item{\code{datf}: the full data set with Y, X, C, Z}
\item{\code{dat0}: the oracle data Y, W, Delta, Z (with no censoring)}
\item{\code{datcc}: the observed data Y, W, Delta, Z}
\item{\code{dat}: the complete cases from the observed data}
}
}
\description{
generate data with X and C having gamma distributions
}
