% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{gen.data}
\alias{gen.data}
\title{generate data}
\usage{
gen.data(n, q, B, s2, x.rate, x.shape, c.rate, c.shape)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{x.rate}{a positive number, rate parameter for gamma distribution of X}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.rate}{a positive number, rate parameter for gamma distribution of C}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}
}
\value{
a data frame
}
\description{
generate data
}
