% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_eq.R
\name{get.a}
\alias{get.a}
\title{compute a function at nodes}
\usage{
get.a(B, s2, mu, d.mu, fy, SF, x.nds, x.wts, c.nds, c.wts, y.wts, y.nds)
}
\arguments{
\item{B}{a numeric vector, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{mu}{a function, mean of outcome Y given covariates X, Z}

\item{fy}{a function, conditional density of outcome Y given covariates X, Z}

\item{SF}{a function, full data score}

\item{x.nds}{a numeric vector, nodes for quadrature rule for X}

\item{x.wts}{a numeric vector, weights for quadrature rule for X}

\item{c.nds}{a numeric vector, nodes for quadrature rule for C}

\item{c.wts}{a numeric vector, weights for quadrature rule for C}

\item{y.wts}{a numeric vector, weights for quadrature rule for Y}

\item{y.nds}{a numeric vector, nodes for quadrature rule for Y}
}
\value{
approximated a values at nodes
}
\description{
compute a function at nodes
}
