% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_beta.R
\name{get.c.param.beta}
\alias{get.c.param.beta}
\title{get rate theta for C given desired censoring proportion}
\usage{
get.c.param.beta(q, x.theta, x.gamma, c.gamma)
}
\arguments{
\item{q}{a number in \link{0,1}, the censoring proportion}
}
\value{
theta parameter for C
}
\description{
get rate theta for C given desired censoring proportion
}
