% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_gamma.R
\name{get.c.rate}
\alias{get.c.rate}
\title{get rate parameter for C given desired right-censoring proportion}
\usage{
get.c.rate(q, x.rate, x.shape, c.shape)
}
\arguments{
\item{q}{a number in \link{0,1}, the right-censoring proportion}

\item{x.rate}{a positive number, rate parameter for gamma distribution of X}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}
}
\value{
rate parameter for gamma distribution of C
}
\description{
get rate parameter for C given desired right-censoring proportion
}
