% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_eq.R
\name{get.d}
\alias{get.d}
\title{compute d vector in integral equation}
\usage{
get.d(x.nds, x.wts, c.nds, c.wts)
}
\arguments{
\item{x.nds}{a numeric vector, nodes for quadrature rule for X}

\item{x.wts}{a numeric vector, weights for quadrature rule for X}

\item{c.nds}{a numeric vector, nodes for quadrature rule for C}

\item{c.wts}{a numeric vector, weights for quadrature rule for C}
}
\value{
d matrix
}
\description{
compute d vector in integral equation
}
