% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_eq.R
\name{get.M}
\alias{get.M}
\title{compute M matrix in integral equation}
\usage{
get.M(B, s2, mu, fy, z, x.nds, x.wts, c.nds, c.wts, y.nds, y.wts)
}
\arguments{
\item{B}{a numeric vector, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{mu}{a function, mean of outcome Y given covariates X, Z}

\item{fy}{a function, conditional density of outcome Y given covariates X, Z}

\item{x.nds}{a numeric vector, nodes for quadrature rule for X}

\item{x.wts}{a numeric vector, weights for quadrature rule for X}

\item{c.nds}{a numeric vector, nodes for quadrature rule for C}

\item{c.wts}{a numeric vector, weights for quadrature rule for C}

\item{y.nds}{a numeric vector, nodes for quadrature rule for Y}

\item{y.wts}{a numeric vector, weights for quadrature rule for Y}
}
\value{
M matrix
}
\description{
compute M matrix in integral equation
}
