% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_beta.R
\name{get.q.beta}
\alias{get.q.beta}
\title{get right-censoring probability q given X and C ~ Beta}
\usage{
get.q.beta(x.theta, x.gamma, c.theta, c.gamma)
}
\arguments{
\item{x.theta}{a positive number, theta parameter for X}

\item{x.gamma}{a positive number, gamma parameter for X}

\item{c.theta}{a positive number, theta parameter for C}

\item{c.gamma}{a positive number, gamma parameter for C}
}
\value{
a number in \link{0,1}, the right-censoring proportion q
}
\description{
get right-censoring probability q given X and C ~ Beta
}
