% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_gamma.R
\name{get.q.gamma}
\alias{get.q.gamma}
\title{get censoring probability q given X and C ~ gamma}
\usage{
get.q.gamma(x.rate, x.shape, c.rate, c.shape)
}
\arguments{
\item{x.rate}{a positive number, rate parameter for gamma distribution of X}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.rate}{a positive number, rate parameter for gamma distribution of C}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}
}
\value{
censoring proportion q
}
\description{
get censoring probability q given X and C ~ gamma
}
