% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_funs.R
\name{get.root}
\alias{get.root}
\title{find root of estimating function}
\usage{
get.root(dat, score, start, args = list())
}
\arguments{
\item{dat}{a data frame including columns Y, W, Delta, Z}

\item{score}{a function of theta = c(beta, log(s2)), the estimating function
to be solved}

\item{start}{a numeric vector, starting value for root search}

\item{args}{list of additional arguments}
}
\value{
root of estimating function
}
\description{
find root of estimating function
}
