% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_eq.R
\name{interp.a}
\alias{interp.a}
\title{interpolate a function between nodes}
\usage{
interp.a(a.vals, x.nds, x.wts, x.new, eta1)
}
\arguments{
\item{a.vals}{a numeric matrix, values of a function at nodes}

\item{x.nds}{a numeric vector, nodes for quadrature rule for X}

\item{x.wts}{a numeric vector, weights for quadrature rule for X}

\item{x.new}{a numeric vector, x values at which to interpolate a}

\item{eta1}{a function, density of X}
}
\value{
interpolated values of a
}
\description{
interpolate a function between nodes
}
