% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_extra.R
\name{interpolate.qtl.power}
\alias{interpolate.qtl.power}
\title{Interpolates QTL mapping power for sample populations with replicates from dense results 
in single observation simulations}
\usage{
interpolate.qtl.power(r1.results, qtl.effect.sizes, strain.effect.sizes = 0,
  num.replicates, n.alleles, use.window = TRUE, n.strains)
}
\arguments{
\item{r1.results}{Data frame of power estimates from simulations based on a single observation per strain.
r1.dat and r1.damb.dat are included in SPARCC for this purpose.}

\item{qtl.effect.sizes}{The desired QTL effect sizes for which power estimates will be interpolated from dense
single observation results.}

\item{strain.effect.sizes}{The desired proportions of variance due to background strain effect for hypothetical population in which
power is evaluated.}

\item{num.replicates}{The desired number of replicates for a hypothetical study in which to evaluate power.}

\item{n.alleles}{The number of functional alleles for the simulated QTL.}

\item{use.window}{DEFAULT: TRUE. Whether the interpolated powers are based on denser power estimates using a window
around a locus or just the actual locus.}

\item{n.strains}{The number of CC strains for the desired interpolated power estimate.}
}
\description{
This function interpolates the QTL mapping power for an experimental mapping population with replicates
based on dense simulated power results in simulations with only a single observation per strain.
}
\examples{
interpolate.qtl.power()
}
