% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_fit.R
\name{inv.softmax}
\alias{inv.softmax}
\title{Inverse softmax transformation}
\usage{
inv.softmax(x, wts)
}
\arguments{
\item{x}{a numeric vector, values to be transformed}

\item{wts}{a numeric vector, weights}
}
\value{
a numeric vector
}
\description{
Inverse softmax transformation
}
