% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_variance_helpers.R
\name{log.fxz}
\alias{log.fxz}
\title{log-likelihood assuming X|Z ~ Beta}
\usage{
\method{log}{fxz}(dat, alpha)
}
\arguments{
\item{dat}{data frame, the observed data}

\item{alpha}{a numeric vector, beta parameters for X|Z}
}
\value{
a matrix of estimating equation values
}
\description{
log-likelihood assuming X|Z ~ Beta
}
