% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{SF}
\alias{SF}
\title{Full data score function}
\usage{
SF(y, x, z, B, ls2)
}
\arguments{
\item{x}{a numeric vector, covariate values}

\item{z}{a numeric vector, covariate values}

\item{B}{a numeric vector, parameter value}

\item{ls2}{a number, the log of the variance of Y|X,Z}
}
\description{
Full data score function
}
