% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim0}
\alias{sim0}
\title{simulation setup 0}
\usage{
sim0(
  n,
  q,
  B2,
  s2,
  x.mean,
  x.shape,
  c.shape,
  specify.x.gamma,
  specify.c.gamma,
  mx = 100,
  mc = 15,
  my = 3,
  seed
)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B2}{a number, the coefficient for \code{X} in the outcome model}

\item{s2}{a number, the variance in the outcome model}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}

\item{specify.x.gamma}{logical, an indicator of whether X should be estimated
as gamma (\code{T}) or exponential (\code{F})}

\item{specify.c.gamma}{logical, an indicator of whether C should be estimated
as gamma (\code{T}) or exponential (\code{F})}

\item{mx}{a positive number, the number of nodes in quadrature grid for X}

\item{mc}{a positive number, the number of nodes in quadrature grid for C}

\item{my}{a positive number, the number of nodes in quadrature grid for Y}

\item{seed}{a positive integer, the seed value for random number generation}
}
\description{
simulation setup 0
}
