% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim1}
\alias{sim1}
\title{simulation setup 1}
\usage{
sim1(
  n,
  q,
  B2,
  s2,
  x.theta.scale,
  x.gamma,
  c.gamma,
  x.correct,
  c.correct,
  mx = 50,
  mc = 15,
  my = 3,
  seed
)

sim1(
  n,
  q,
  B2,
  s2,
  x.theta.scale,
  x.gamma,
  c.gamma,
  x.correct,
  c.correct,
  mx = 50,
  mc = 15,
  my = 3,
  seed
)
}
\arguments{
\item{B2}{a number, the coefficient for \code{X} in the outcome model}

\item{s2}{a number, the variance in the outcome model}

\item{x.correct}{logical, an indicator of whether X should be estimated
as beta (\code{T}) or uniform (\code{F})}

\item{c.correct}{logical, an indicator of whether C should be estimated
as beta (\code{T}) or uniform (\code{F})}

\item{mx}{a positive number, the number of nodes in quadrature grid for X}

\item{mc}{a positive number, the number of nodes in quadrature grid for C}

\item{my}{a positive number, the number of nodes in quadrature grid for Y}

\item{seed}{a positive integer, the seed value for random number generation}

\item{specify.x.gamma}{logical, an indicator of whether X should be estimated
as gamma (\code{T}) or exponential (\code{F})}

\item{specify.c.gamma}{logical, an indicator of whether C should be estimated
as gamma (\code{T}) or exponential (\code{F})}
}
\description{
simulation setup 1

simulation setup 2
}
