% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{sim2}
\alias{sim2}
\title{simulation 2: assess methods over sequence of censoring proportion}
\usage{
sim2(
  n,
  q,
  B2,
  s2,
  x.theta.scale,
  x.gamma,
  c.gamma,
  mx = 40,
  mc = 40,
  my = 4,
  seed
)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B2}{a number, the coefficient for \code{X} in the outcome model}

\item{s2}{a number, the variance in the outcome model}

\item{x.gamma}{a positive number, gamma parameter for X}

\item{c.gamma}{a positive number, gamma parameter for C}

\item{mx}{a positive number, the number of nodes in quadrature grid for X}

\item{mc}{a positive number, the number of nodes in quadrature grid for C}

\item{my}{a positive number, the number of nodes in quadrature grid for Y}

\item{seed}{a positive integer, the seed value for random number generation}
}
\description{
simulation 2: assess methods over sequence of censoring proportion
}
