% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation3.R
\name{sim3}
\alias{sim3}
\title{simulation 1: assess various methods with 3D Z}
\usage{
sim3(n, q, ndistinct.Z, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{ndistinct.Z}{a positive integer, the number of distinct Z values}

\item{seed}{a positive integer, the seed value for random number generation}
}
\description{
simulation 1: assess various methods with 3D Z
}
