% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation4.R
\name{sim4}
\alias{sim4}
\title{simulation 4: assess the sparcc estimator with poor B-spline performance}
\usage{
sim4(n, q, m.knots, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{m.knots}{a positive integer, the number of B-spline knots}

\item{seed}{a positive integer, the seed value for random number generation}
}
\description{
simulation 4: assess the sparcc estimator with poor B-spline performance
}
