% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{sparcc}
\alias{sparcc}
\title{Fit SPARCC Estimator}
\usage{
sparcc(
  data,
  xz.interaction = T,
  nuisance.models = "parametric",
  distr.x = "beta",
  distr.c = "beta",
  m.knots = 5,
  deg = 3,
  mx = 40,
  my = 40,
  mc = 5,
  range.x = c(1e-06, 1 - 1e-06),
  range.c = c(1e-06, 1 - 1e-06)
)
}
\arguments{
\item{data}{a data frame with the following columns
\itemize{
\item{\code{Y}: outcome}
\item{\code{W}: minimum of the censored covariate \code{X} and the censoring time \code{C}}
\item{\code{Delta}: an indicator if \code{X} is observed (\code{Delta = 1}) or censored
(\code{Delta = 0})}
\item{\code{Z}: a discrete-valued uncensored covariate}
}}

\item{xz.interaction}{a logical indicator for whether an \code{X*Z} interaction is
to be included in the outcome model (\code{TRUE}) or not (\code{FALSE}), default is
\code{TRUE}}

\item{nuisance.models}{model type for nuisance distribution models
("parametric" or "nonparametric)}

\item{distr.x}{a character string "name" naming a distribution for \code{X|Z},
for which the corresponding density function dname must be defined; used only
if \code{nuisance.models} is "parametric"}

\item{distr.c}{a character string "name" naming a distribution for \code{C|Z},
for which the corresponding density function dname must be defined; used only
if \code{nuisance.models} is "parametric"}

\item{m.knots}{a positive integer, the number of knots in the nonparametric
B-spline estimator for the densities of of \code{X|Z} and \code{C|Z}, default is 5;
used only if \code{nuisance.models} is "nonparametric"}

\item{deg}{a positive integer, the degree of polynomials in the nonparametric
B-spline estimator for the densities of of \code{X|Z} and \code{C|Z}, default is 3;
used only if \code{nuisance.models} is "nonparametric"}

\item{mx}{a positive integer, the number of quadrature nodes used for \code{X},
default is 40}

\item{my}{a positive integer, the number of Gauss-Hermite quadrature nodes
used for \code{Y}, default is 5}

\item{mc}{a positive integer, the number of quadrature nodes used for \code{C},
default is 40}

\item{range.x}{a numeric vector of length 2, the range of the support of \code{X},
default is c(1E-6, 1-1E-6)}

\item{range.c}{a numeric vector of length 2, the range of the support of \code{C},
default is c(1E-6, 1-1E-6)}
}
\value{
a list of arguments including
\itemize{
\item{\code{x.model}: a list with
\code{x.model.type} (parametric or nonparametric),
\code{distr.x}, \code{m.knots}, \code{deg},
\code{x.params.hat} (the estimated parameters for the distribution of \code{X|Z} at
each level of \code{Z}),
\code{x.nds} (the quadrature nodes used for \code{X}),
\code{x.wts} (the quadrature weights used for \code{X}),
\code{eta1} (the estimated density function of \code{X|Z} at specific nodes)
}
\item{\code{c.model}: a list with
\code{c.model.type} (parametric or nonparametric),
\code{distr.c}, \code{m.knots}, \code{deg},
\code{c.params.hat} (the estimated parameters for the distribution of \code{C|Z} at
each level of \code{Z}),
\code{c.nds} (the quadrature nodes used for \code{C}),
\code{c.wts} (the quadrature weights used for \code{C}),
\code{eta2} (the estimated density function of \code{C|Z} at specific nodes)
}
\item{\code{outcome.model}: a list with
\code{outcome.fmla} (the outcome model formula),
\code{coef} (the outcome model coefficient estimates),
\code{cov} (the estimated covariance matrix of the coefficient estimator)}
}
}
\description{
To fit the SPARCC estimator for linear model with a conditionally normal
outcome \code{Y}, a censored covariate \code{X}, and a discrete-valued uncensored
covariate \code{Z}.
}
