% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_est.R
\name{var.est}
\alias{var.est}
\title{influence function based variance estimator}
\usage{
var.est(dat, theta, args, get.S, return.se = F)
}
\arguments{
\item{dat}{data frame, the observed data}

\item{theta}{a numeric vector, outcome parameters (beta, log(s2))}

\item{args}{a list of additional arguments}

\item{get.S}{a function of theta, the estimating function}
}
\value{
.se an optional logical indicator for returning a vector of standard
errors as opposed to a covariance matrix, defaults to FALSE

a covariance matrix or vector of standard errors
}
\description{
influence function based variance estimator
}
