% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_est.R
\name{var.est.sand}
\alias{var.est.sand}
\title{sandwich variance estimator}
\usage{
var.est.sand(dat, theta, args, get.S, return.se = F)
}
\arguments{
\item{dat}{data frame, the observed data}

\item{theta}{a numeric vector, outcome parameters (beta, log(s2))}

\item{args}{a list of additional arguments}

\item{get.S}{a function of theta, the estimating function}
}
\value{
a covariance matrix or vector of standard errors
}
\description{
sandwich variance estimator
}
