% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.dat.beta}
\alias{assess.dat.beta}
\title{Assess data generation with X|Z and C|Z having beta distributions}
\usage{
assess.dat.beta(n, q, B, s2, x.thetas, x.gamma, c.gamma, x.correct, c.correct)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{B}{a vector of numbers, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{x.gamma}{a positive number, gamma parameter for X}

\item{c.gamma}{a positive number, gamma parameter for C}

\item{x.mean}{a positive number, the mean of X}
}
\value{
a plot of generated data
}
\description{
Assess data generation with X|Z and C|Z having beta distributions
}
