% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{assess.dat.gamma}
\alias{assess.dat.gamma}
\title{Assess data generation with X|Z and C|Z having gamma distributions}
\usage{
assess.dat.gamma(
  n,
  q,
  B,
  s2,
  x.means,
  x.shape,
  c.shape,
  specify.x.gamma,
  specify.c.gamma
)
}
\arguments{
\item{x.mean}{a positive number, the mean of X}
}
\value{
a plot of generated data
}
\description{
Assess data generation with X|Z and C|Z having gamma distributions
}
