% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{d.mu}
\alias{d.mu}
\title{gradient of linear mean function mu w.r.t. B}
\usage{
d.mu(x, z, B, xz.interaction)
}
\arguments{
\item{x}{a numeric vector, covariate values}

\item{z}{a numeric vector, covariate values}

\item{B}{a numeric vector, parameter value}

\item{xz.interaction}{logical indicator for including X*Z interaction}
}
\value{
a numeric matrix, the gradient of mu w.r.t. B
}
\description{
gradient of linear mean function mu w.r.t. B
}
