% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{fy}
\alias{fy}
\title{normal density for Y|X,Z}
\usage{
fy(y, x, z, B, s2, xz.interaction)
}
\arguments{
\item{y}{a numeric vector, outcome values}

\item{x}{a numeric vector, covariate values}

\item{z}{a numeric vector, covariate values}

\item{B}{a numeric vector, parameter value}

\item{s2}{a positive number, the variance of Y|X,Z}

\item{xz.interaction}{logical indicator for including X*Z interaction}
}
\description{
normal density for Y|X,Z
}
