% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_beta.R
\name{get.c.param.beta}
\alias{get.c.param.beta}
\title{get beta parameters for C given desired censoring proportion}
\usage{
get.c.param.beta(q, x.theta, x.gamma, c.gamma)
}
\arguments{
\item{q}{a number in \link{0,1}, the censoring proportion}

\item{x.theta}{a positive number, theta parameter for X}

\item{x.gamma}{a positive number, gamma parameter for X}

\item{c.gamma}{a positive number, gamma parameter for C}
}
\value{
a positive number, the theta parameter for C
}
\description{
get beta parameters for C given desired censoring proportion
}
