% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_gamma.R
\name{get.c.rate}
\alias{get.c.rate}
\title{get rate parameter for C given desired censoring proportion}
\usage{
get.c.rate(q, x.rate, x.shape, c.shape)
}
\arguments{
\item{q}{a number in \link{0,1}, the censoring proportion}
}
\value{
rate parameter for gamma distribution of C
}
\description{
get rate parameter for C given desired censoring proportion
}
