% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_funs.R
\name{get.Scc}
\alias{get.Scc}
\title{complete case score function}
\usage{
get.Scc(dat, B, s2, args, return.sums = T)
}
\arguments{
\item{dat}{a data frame including columns Y, W, Delta, Z}

\item{B}{a numeric vector, parameters in the outcome model}

\item{s2}{a positive number, variance in the outcome model}

\item{args}{list of additional arguments}

\item{return.sums}{logical indicator for returning sum of scores as opposed
to individual scores, default is TRUE}
}
\value{
complete case score function values
}
\description{
complete case score function
}
