% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_funs.R
\name{mu}
\alias{mu}
\title{linear mean function mu(X, Z, B) = E(Y | X, Z)}
\usage{
mu(x, z, B, xz.interaction)
}
\arguments{
\item{x}{a numeric vector, covariate values}

\item{z}{a numeric vector, covariate values}

\item{B}{a numeric vector, parameter value}

\item{xz.interaction}{logical indicator for including X*Z interaction}
}
\value{
a numeric vector, mean values
}
\description{
linear mean function mu(X, Z, B) = E(Y | X, Z)
}
