% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim1}
\alias{sim1}
\title{simulation setup 1}
\usage{
sim1(n, q, x.shape = 1, c.shape = 1, mx = 50, mc = 15, my = 2, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{x.shape}{a positive number, shape parameter for gamma distribution of X}

\item{c.shape}{a positive number, shape parameter for gamma distribution of C}

\item{mx}{a positive number, the number of nodes in quadrature grid for X}

\item{mc}{a positive number, the number of nodes in quadrature grid for C}

\item{my}{a positive number, the number of nodes in quadrature grid for Y}

\item{seed}{a positive integer, the seed value for random number generation}
}
\description{
simulation setup 1
}
