% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation1.R
\name{sim1}
\alias{sim1}
\title{simulation 1: assess various methods under misspecification}
\usage{
sim1(n, q, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{q}{a number in \link{0,1}, the censoring proportion}

\item{seed}{a positive integer, the seed value for random number generation}

\item{B2}{a number, the coefficient for \code{X} in the outcome model}

\item{s2}{a number, the variance in the outcome model}

\item{mx}{a positive number, the number of nodes in quadrature grid for X}

\item{mc}{a positive number, the number of nodes in quadrature grid for C}

\item{my}{a positive number, the number of nodes in quadrature grid for Y}
}
\description{
simulation 1: assess various methods under misspecification
}
